requirements.txt
asgiref==3.5.2
bcrypt==3.1.7
dj-database-url==1.0.0
Django==4.1.1
djangorestframework==3.14.0
gunicorn==20.1.0
httplib2==0.14.0
oauthlib==3.1.0
pandas==1.4.4
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1 
whitenoise==6.2.0
django-cors-headers==3.13.0

